// src/app/modules/EnrolledCourse/enrolledCourse.interface.ts
import { Types } from 'mongoose'

export type TGrade = 'A' | 'B' | 'C' | 'D' | 'F' | 'NA'

export type TEnrolledCourseMarks = {
  classTest1: number
  midTerm: number
  classTest2: number
  finalTerm: number
}

export type TEnrolledCourse = {
  transaction: Types.ObjectId
  academicDepartment: Types.ObjectId
  offeredCourse: Types.ObjectId
  course: Types.ObjectId
  student: Types.ObjectId
  faculty: Types.ObjectId
  isEnrolled: boolean
  courseMarks: TEnrolledCourseMarks
  grade: TGrade
  gradePoints: number
  isCompleted: boolean
}